fix: verification de connexion et role pour partie quiz et questions

+++ kurso-e-learning/src/actions/quiz.actions.ts
@@ -1,30 +1,39 @@
 "use server"
 
-import { GET } from "@/app/api/questions/route"
 import prisma from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
+import { getCurrentUser } from "./user.actions";
 
-
-//!!!!!!!!!!!!!!!!!!!!! rewst de verifier l'utilisateur == FORMATEUR
 
 export async function createQuiz(formData: FormData) {
 
+    //api
     const response = await fetch("http://localhost:3000/api/questions");
     if (!response.ok) {
-        throw new Error("Erreur lors de la récupération des questions");
+        throw new Error("Erreur lors de la recuperation des questions");
     }
     const data = await response.json();
+
+    //verification connexion et role
+    const utilisateur = await getCurrentUser();
+    if (!utilisateur) throw new Error("Vous n'etes pas connecter");
+    if(utilisateur.role == "ETUDIANT") throw new Error("Vous n'est pas un formateur, faites une demende pour changer acces.");
     
 
     const titre = formData.get("titre") as string;
     const leconId = formData.get("leconId") as string;
 
+    // verifder la lecon
     const lecon = await prisma.lecon.findUnique({
         where: { id: leconId },
         include: { cours: true },
     });
-    if (!lecon) throw new Error("Leçon introuvable.");
+    if (!lecon) throw new Error("Lecon introuvable.");
+
+
+    //verifier le formateur du cours = utilisateur
+    if (lecon.cours.formateurId !== utilisateur.id) throw new Error("Vous ne pouvez pas creer un quiz dans ce cours");
 
     const quiz = await prisma.quiz.create({
         data:{
@@ -44,7 +53,7 @@ export async function createQuiz(formData: FormData) {
         include: {
             questions: true,
         }
-    })
+    });
 
 
     revalidatePath(`/cours/${lecon.coursId}/lecons/${leconId}`);
@@ -77,9 +86,15 @@ export async function getQuiz(id : string) {
 
 export async function updateQuiz(formData: FormData) {
 
+    const utilisateur = await getCurrentUser();
+    if (!utilisateur) throw new Error("Vous n'etes pas connecter");
+    if(utilisateur.role == "ETUDIANT") throw new Error("Vous n'est pas un formateur, faites une demende pour changer acces.");
+    
+
     const id = formData.get("id") as string; 
     const titre = formData.get("titre") as string;
 
+    //verfier que le cours existe
     const quiz = await prisma.quiz.findUnique({
         where: { id },
         include: { 
@@ -88,8 +103,10 @@ export async function updateQuiz(formData: FormData) {
             }
         },
     });
-
     if (!quiz) throw new Error("Quiz introuvable.");
+
+    //verifier le formateur du cours = utilisateur
+    if (quiz.lecon.cours.formateurId !== utilisateur.id) throw new Error("Vous ne pouvez pas modifier ce quiz.")
 
     await prisma.quiz.update({ 
         where: { 
@@ -101,7 +118,7 @@ export async function updateQuiz(formData: FormData) {
     }); 
     
     revalidatePath(`/cours/${quiz.lecon.coursId}/lecons/${quiz.leconId}/quiz/${id}`); 
-    redirect(`/cours/${quiz.lecon.coursId}/lecons/${quiz.leconId}`);
+    redirect(`/cours/${quiz.lecon.coursId}/lecons/${quiz.leconId}/quiz/${id}`);
 
 
 }
@@ -109,6 +126,11 @@ export async function updateQuiz(formData: FormData) {
 
 
 export async function deleteQuiz(formData: FormData) {
+
+    const utilisateur = await getCurrentUser();
+    if (!utilisateur) throw new Error("Vous n'etes pas connecter");
+    if(utilisateur.role == "ETUDIANT") throw new Error("Vous n'est pas un formateur, faites une demende pour changer acces.");
+
     const id = formData.get("id") as string; 
 
     const quiz = await prisma.quiz.findUnique({
@@ -122,13 +144,17 @@ export async function deleteQuiz(formData: FormData) {
 
     if (!quiz) throw new Error("Quiz introuvable.");
 
+    if (quiz.lecon.cours.formateurId !== utilisateur.id) {
+        throw new Error("Vous ne pouvez pas supprimer ce quiz.");
+    }
+
     await prisma.quiz.delete({ 
         where: { 
             id: id, 
         }, 
     }); 
     
-    revalidatePath(`/cours/${quiz.lecon.coursId}/lecons/${quiz.leconId}`); 
-    redirect(`/cours/${quiz.lecon.coursId}/lecons/${quiz.leconId}`);
+    revalidatePath(`/cours/${quiz.lecon.coursId}/lecons/${quiz.leconId}/quiz`); 
+    redirect(`/cours/${quiz.lecon.coursId}/lecons/${quiz.leconId}/quiz`);
 
 }
